requirements.txt
python-telegram-bot==20.3
boto3==1.28.0
psycopg2-binary>=2.9.9
python-dotenv==1.0.0
pytz==2023.3
schedule==1.2.0